[KAN-68] refactor: 모임 만들기 모달 분리 (종합)
