--- src/app/components/BoxSelect/BoxSelectGroup.tsx
@@ -1,49 +1,0 @@
-'use client';
-
-import BoxSelect from './BoxSelect';
-
-interface GatheringType {
-  [key: string]: boolean;
-}
-
-interface BoxSelectGroupProps {
-  gatheringType: GatheringType;
-  setGatheringType: (type: (prevType: GatheringType) => GatheringType) => void;
-}
-
-const BoxSelectGroup = ({
-  gatheringType,
-  setGatheringType,
-}: BoxSelectGroupProps) => {
-  const handleCheckboxChange = (type: string) => {
-    setGatheringType(() => ({
-      OFFICE_STRETCHING: type === 'OFFICE_STRETCHING',
-      MINDFULLNESS: type === 'MINDFULLNESS',
-      WORKATION: type === 'WORKATION',
-    }));
-  };
-
-  return (
-    <div className='flex gap-12'>
-      <BoxSelect
-        title='달램핏'
-        subTitle='오피스 트레이닝'
-        isSelected={gatheringType.OFFICE_STRETCHING}
-        onChange={() => handleCheckboxChange('OFFICE_STRETCHING')}
-      />
-      <BoxSelect
-        title='달램핏'
-        subTitle='마인드풀니스'
-        isSelected={gatheringType.MINDFULLNESS}
-        onChange={() => handleCheckboxChange('MINDFULLNESS')}
-      />
-      <BoxSelect
-        title='워케이션'
-        isSelected={gatheringType.WORKATION}
-        onChange={() => handleCheckboxChange('WORKATION')}
-      />
-    </div>
-  );
-};
-
-export default BoxSelectGroup;
