+++ src/app/components/Modal/MakeGatheringModal/ImageUploader.tsx
@@ -20,23 +20,26 @@ const ImageUploader = ({ image, setImage }: ImageUploaderProps) => {
   };
 
   return (
-    <div>
-      <input
-        ref={inputRef}
-        type='file'
-        className='hidden'
-        onChange={handleChange}
-      />
-      <div className='flex gap-12'>
-        <div className='flex w-full items-center rounded-xl bg-gray-50 px-16 py-[10px]'>
-          {image ?? <p className='text-gray-400'>이미지를 첨부해 주세요</p>}
-        </div>
-        <div className='w-100'>
-          <Button
-            name='파일 찾기'
-            variant='white'
-            onClick={() => inputRef.current?.click()}
-          />
+    <div className='space-y-12 text-16 font-semibold'>
+      <h2>이미지</h2>
+      <div>
+        <input
+          ref={inputRef}
+          type='file'
+          className='hidden'
+          onChange={handleChange}
+        />
+        <div className='flex gap-12'>
+          <div className='flex w-full items-center rounded-xl bg-gray-50 px-16 py-[10px]'>
+            {image ?? <p className='text-gray-400'>이미지를 첨부해 주세요</p>}
+          </div>
+          <div className='w-100'>
+            <Button
+              name='파일 찾기'
+              variant='white'
+              onClick={() => inputRef.current?.click()}
+            />
+          </div>
         </div>
       </div>
     </div>
