+++ src/app/components/Modal/MakeGatheringModal/PlaceDropdown.tsx
@@ -65,12 +65,15 @@ const PlaceDropdown = ({
 
   return (
     <div className='relative' ref={containerRef}>
-      <div
-        className={`h- mb-8 flex w-full cursor-pointer items-center justify-between rounded-xl py-[6px] pl-12 pr-8 text-16 font-semibold md:h-44 md:py-[10px] ${stateClasses[currentState]}`}
-        onClick={toggleDropDown}
-      >
-        {selectedOption || children}
-        <IconCaret className={`h-24 w-24 text-var-gray-800`} />
+      <div className='space-y-12 text-16 font-semibold'>
+        <h2>장소</h2>
+        <div
+          className={`h- mb-8 flex w-full cursor-pointer items-center justify-between rounded-xl py-[6px] pl-12 pr-8 text-16 font-semibold md:h-44 md:py-[10px] ${stateClasses[currentState]}`}
+          onClick={toggleDropDown}
+        >
+          {selectedOption || children}
+          <IconCaret className={`h-24 w-24 text-var-gray-800`} />
+        </div>
       </div>
 
       {isOpen && (
